% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinate_Converter.R
\name{Coordinate_Covnerter1}
\alias{Coordinate_Covnerter1}
\title{Coordinate Converter 1}
\usage{
Coordinate_Covnerter1(
  chromosome,
  genomic_start,
  genomic_end,
  refAllele,
  varAllele,
  strand
)
}
\arguments{
\item{chromosome}{sequence name like "chrX"}

\item{genomic_start}{genomic start position}

\item{genomic_end}{genomic end position}

\item{refAllele}{reference allele}

\item{varAllele}{variant allele}

\item{strand}{sequence strand "+" or "-"}
}
\value{
dataframe which contians corresponding HUGO Symbol, transcript ID, position in primary transcript, exchanged amino acid.
}
\description{
This function take genomic location,  nucleotide exchange as input and convert these to a dataframe which contains
HUGO Symbol, transcript ID, position in primary transcript, exchanged amino acid.
}
\examples{
DF1 <- Coordinate_Covnerter1("chrX",48823056,48823056,"G","C")

}
