% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinate_Converter.R
\name{Coordinate_Covnerter2}
\alias{Coordinate_Covnerter2}
\title{Coordinate Converter 2}
\usage{
Coordinate_Covnerter2(
  chromosome,
  genomic_start,
  genomic_end,
  refAllele,
  varAllele,
  strand
)
}
\arguments{
\item{chromosome}{sequence name like "chrX"}

\item{genomic_start}{genomic start position}

\item{genomic_end}{genomic end position}

\item{refAllele}{reference allele}

\item{varAllele}{variant allele}

\item{strand}{sequence strand "+" or "-"}
}
\value{
dataframe which contians corresponding HUGO Symbol, transcript ID, position exchanged amino acid.
}
\description{
This function take genomic location,  nucleotide exchange as input and convert these to a dataframe which contains
HUGO Symbol, transcript ID, position exchanged amino acid.
}
