% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinate_Converter.R
\name{Coordinate_Covnerter3}
\alias{Coordinate_Covnerter3}
\title{Coordinate Converter 3}
\usage{
Coordinate_Covnerter3(chromosome, genomic_start, genomic_end)
}
\arguments{
\item{chromosome}{sequence name like "chrX"}

\item{genomic_start}{genomic start position}

\item{genomic_end}{genomic end position}
}
\value{
dataframe which contians corresponding HUGO Symbol, list of all transcript ID.
}
\description{
This function take genomic location as input and convert these to a dataframe which contains
corresponding HUGO Symbol, list of all transcript ID.
}
