% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinate_Converter.R
\name{Coordinate_Covnerter5_1}
\alias{Coordinate_Covnerter5_1}
\title{Coordinate Converter 5_1}
\usage{
Coordinate_Covnerter5_1(transcript_id, TX_start, TX_end)
}
\arguments{
\item{transcript_id}{transcript passed in should be external transcript name like "HDAC6-201"}

\item{TX_start}{transcript start position like 2657}

\item{TX_end}{transcript end position like 2657}
}
\value{
dataframe which contians corresponding genomic position, codon and codon's number.
}
\description{
This function take transcript id and transcript location as input and convert these to a dataframe which contains
corresponding genomic position, codon and codon's number.
}
\examples{
DF5_1<-Coordinate_Covnerter5_1("HDAC6-201",2657,2657)
}
