% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinate_Converter.R
\name{Coordinate_Covnerter5_2}
\alias{Coordinate_Covnerter5_2}
\title{Coordinate Converter 5_2}
\usage{
Coordinate_Covnerter5_2(transcript_id, TX_position)
}
\arguments{
\item{transcript_id}{transcript id passed in should be ensembl transcript id like "ENST00000334136"}

\item{TX_position}{transcript position like 2657}
}
\value{
dataframe which contians corresponding genomic position, codon and codon's number.
}
\description{
This function take transcript id and transcript location as input and convert these to a dataframe which contains
corresponding genomic position, codon and codon's number.
}
