% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinate_Converter.R
\name{Coordinate_Covnerter6}
\alias{Coordinate_Covnerter6}
\title{Coordinate Converter 6}
\usage{
Coordinate_Covnerter6(
  chromosome,
  genomic_start,
  genomic_end,
  genomeVersion,
  targetGenomeVersion
)
}
\arguments{
\item{chromosome}{sequence name like "chrX"}

\item{genomic_start}{genomic start position}

\item{genomic_end}{genomic end position}

\item{genomeVersion}{original genome version}

\item{targetGenomeVersion}{target genome version which want to transfer}
}
\value{
dataframe which contians corresponding converted target genome coordinate.
}
\description{
This function take genomic location, orginal and target genome version as input and convert these to a dataframe which contains
corresponding genomic position.
}
\examples{
DF6 <- Coordinate_Covnerter6("chrX",45060024,45060024,"hg16","hg18")
}
