% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunction.R
\name{Coordinate_Mapping_gffTB}
\alias{Coordinate_Mapping_gffTB}
\title{Coordinate select function for gff dataframe Table.}
\usage{
Coordinate_Mapping_gffTB(chr, start_pos, end_pos, gff_Table)
}
\arguments{
\item{chr}{sequence name like "chrX"}

\item{start_pos}{genomic start position}

\item{end_pos}{genomic end position}

\item{gff_Table}{gff data frame generated from gff file}
}
\value{
selected dataframe according to chr, genomic start and end coordinate.
}
\description{
Coordinate select function for gff dataframe Table.
}
