% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunction.R
\name{Coordinate_to_codingTranscriptTB}
\alias{Coordinate_to_codingTranscriptTB}
\title{Coordinate coding table to full transcript coding table}
\usage{
Coordinate_to_codingTranscriptTB(CodingTable, gff_DataTable)
}
\arguments{
\item{CodingTable}{AA coding table generated from predictCoding()}

\item{gff_DataTable}{gff table generated from gff file}
}
\value{
full merged transcript-coding table
}
\description{
Coordinate coding table to full transcript coding table
}
