% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunction.R
\name{Coordinate_to_codons}
\alias{Coordinate_to_codons}
\title{Coordinate to codons table}
\usage{
Coordinate_to_codons(chr, start_pos, end_pos, GFFCDSTable)
}
\arguments{
\item{chr}{sequence name like "chrX"}

\item{start_pos}{genomic start position}

\item{end_pos}{genomic end position}

\item{GFFCDSTable}{cds_gff table extracted from gff table.}
}
\value{
table with coorespoding codons and codons number
}
\description{
Coordinate to codons table
}
