% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunction.R
\name{Coordinate_to_transcriptTB}
\alias{Coordinate_to_transcriptTB}
\title{Coordinate coding table to transcript coding table}
\usage{
Coordinate_to_transcriptTB(CodingTable, GFFCDSTable)
}
\arguments{
\item{CodingTable}{AA coding table generated from predictCoding()}

\item{GFFCDSTable}{cds_gff table extracted from gff table.}
}
\value{
coding table with corresponding transcript id
}
\description{
Coordinate coding table to transcript coding table
}
