% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunction.R
\name{download_GFF_fromURI}
\alias{download_GFF_fromURI}
\title{Download gff file from URI}
\usage{
download_GFF_fromURI(GFFURI)
}
\arguments{
\item{GFFURI}{the gff file URI link}
}
\value{
gff file downloaded path.
}
\description{
Download gff file from URI
}
