% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunction.R
\name{GFFDataload}
\alias{GFFDataload}
\title{Generate TxDB, gff dataframe, cds_gff dataframe based on gff file}
\usage{
GFFDataload(gff_File)
}
\arguments{
\item{gff_File}{gff file located path}
}
\value{

}
\description{
Generate TxDB, gff dataframe, cds_gff dataframe based on gff file
}
\examples{
GFFDataload("~/Downloads/GIT/Tmp_Data/Homo_sapiens.GRCh38.99.gff3.gz")
}
